+++ src/dnds/visualizer.py
@@ -16,6 +16,8 @@ class Visualizer:
     ORG_CLASS_IDX = 0
     ORG_NAME_IDX = 1
 
+    UNACCEPTED_CHARS = ["W", "S", "M", "K", "R", "Y", "B", "D", "H", "V", "N", "Z"]
+
     def __init__(self):
         """ Constructor """
         self.alignments = self._parse_organisms_alignments()
@@ -38,6 +40,8 @@ class Visualizer:
                 header += "\n"
                 sof.write(header)
                 for org in self.scores[sig_org].keys():
+                    if sig_org == org:  # no point in adding the same one
+                        continue
                     vals = ",".join([str(x) for x in self.scores[sig_org][org].values()])
                     line = "{},{}\n".format(org, vals)
                     sof.write(line)
@@ -120,18 +124,23 @@ class Visualizer:
                             # in this case, one of the organisms does not have a gene
                             # assume the score is 1 for neutrality
                             continue
-                        seq1_clean, seq2_clean = self._clean_sequence(seq1, seq2)
-                        try:
-                            score = self.get_dnds(seq1_clean, seq2_clean)[0]  # results come back as vectors
-                            # placing a 1 for neutrality when we get NaNs, based on docs, the sequences should be
-                            # highly divergent but don't know for sure, safer to place in neutral, might change
-                            # mind though...
-                            final[org1][org2][gene] = round(score, 2) if not np.isnan(score) else 1
-                        except Exception:
-                            final[org1][org2][gene] = 0
+                        seq1_clean, seq2_clean = self._clean_sequence(seq1, seq2,
+                                                                      check_chars=True)
+                        # apparently, it can happen that we get non-unique sequences after cleaning
+                        if seq1_clean == seq2_clean:
+                            final[org1][org2][gene] = 1
+                        else:
+                            try:
+                                score = self.get_dnds(seq1_clean, seq2_clean)[0]  # results come back as vectors
+                                # placing a 1 for neutrality when we get NaNs, based on docs, the sequences should be
+                                # highly divergent but don't know for sure, safer to place in neutral, might change
+                                # mind though...
+                                final[org1][org2][gene] = round(score, 2) if not np.isnan(score) else 1
+                            except Exception:
+                                final[org1][org2][gene] = 0
         return final
 
-    def _clean_sequence(self, seq1, seq2):
+    def _clean_sequence(self, seq1, seq2, check_chars=False):
         """
         Cleans the given pair of sequences such that dnds can be computed
         :param str seq1: first sequence to clean
@@ -141,7 +150,13 @@ class Visualizer:
         l_seq1 = list(seq1)
         l_seq2 = list(seq2)
         for i in range(len(l_seq1)):  # or whatever
-            if l_seq1[i] == "-" or l_seq2[i] == "-":
+            illegal = l_seq1[i] == "-" or l_seq2[i] == "-"
+            if check_chars:
+                # there are sequences that contain non-IUPAC char, which cannot be used for dnds evaluation
+                illegal = self._is_illegal_char(l_seq1[i]) or self._is_illegal_char(l_seq2[i])
+                if illegal:
+                    l_seq1[i], l_seq2[i] = "@", "@"
+            elif illegal:
                 # mark characters for removal, this is necessary as strings are immutable
                 l_seq1[i], l_seq2[i] = "@", "@"
         seq1, seq2 = "".join(l_seq1).replace("@", ""), "".join(l_seq2).replace("@", "")
@@ -153,11 +168,10 @@ class Visualizer:
         seq1_remainder = len(seq1_trimmed) % 3
         if not seq1_remainder == 0:
             seq1_trimmed = seq1_trimmed[:len(seq1_trimmed) - seq1_remainder]
-
-        # these should be equal remainder-wise...
         seq2_remainder = len(seq2_trimmed) % 3
         if not seq2_remainder == 0:
             seq2_trimmed = seq2_trimmed[:len(seq2_trimmed) - seq2_remainder]
+        assert (len(seq1_trimmed) == len(seq2_trimmed))
         return seq1_trimmed, seq2_trimmed
 
     def visualize(self):
@@ -186,6 +200,21 @@ class Visualizer:
             plt.savefig(save_fig_path,
                         quality=95,
                         bbox_inches='tight')
+            plt.clf()
+
+    def _contains_illegal_chars(self, seq):
+        """ Check whether a given sequence contains non-ATCG letters (IUPAC accepted but cannot compute dn/ds with
+        those) """
+        for u in self.UNACCEPTED_CHARS:
+            if u in seq:
+                return True
+        return False
+
+    def _is_illegal_char(self, c):
+        """ Check if a character is illegal - non-IUPAC nucleotides """
+        if c in self.UNACCEPTED_CHARS:
+            return True
+        return False
 
 
 if __name__ == "__main__":